refactor(daily-sync): decompose DailySync into 7 mixin modules
Extract 20 methods from 2159-line daily_sync.py into 7 mixin modules
under workflows/sync/, following the same pattern as WorkflowCoach (PR #52):

- activity_tracker.py: ActivityTracker standalone class (5 methods)
- activity_detection.py: ActivityDetectionMixin (3 methods)
- ai_analysis.py: AIAnalysisMixin (2 methods)
- servo_evaluation.py: ServoEvaluationMixin (4 methods)
- ctl_peaks.py: CTLPeaksMixin (1 method)
- session_updates.py: SessionUpdatesMixin (4 methods)
- reporting.py: ReportingMixin (3 methods)

Facade remains at 448 lines with __init__, run(), main(), and
calculate_current_week_info(). All backward-compatible imports preserved.
Updated ~19 test patch paths across test_daily_sync.py and
test_mcp_extra_handlers.py. 1809 tests pass, 15/15 pre-commit hooks.
