+++ magma_cycling/workflows/sync/__init__.py
@@ -1,0 +1,1 @@
+"""DailySync mixin modules for god class decomposition."""
